fix: make sure that the vertex implementation is exported only for the node build
