+++ packages/core/e2e/examples/cloudflare-worker-agent/wrangler.toml
@@ -3,7 +3,6 @@ name = "agent"
 main = "src/index.ts"
 compatibility_date = "2024-04-23"
 compatibility_flags = ["nodejs_compat"]
-nodejs_compat = true
 
 # Automatically place your workloads in an optimal location to minimize latency.
 # If you are running back-end logic in a Worker, running it closer to your back-end infrastructure
