+++ packages/core/src/embeddings/GeminiEmbedding.ts
@@ -1,8 +1,5 @@
-import {
-  GEMINI_BACKENDS,
-  GeminiSessionStore,
-  type GeminiSession,
-} from "../llm/gemini.js";
+import { GeminiSession, GeminiSessionStore } from "../llm/gemini/base.js";
+import { GEMINI_BACKENDS } from "../llm/gemini/types.js";
 import { BaseEmbedding } from "./types.js";
 
 export enum GEMINI_EMBEDDING_MODEL {
