+++ packages/core/src/llm/index.ts
@@ -5,14 +5,13 @@ export {
   Anthropic,
 } from "./anthropic.js";
 export { FireworksLLM } from "./fireworks.js";
+export { Gemini, GeminiSession } from "./gemini/base.js";
+
 export {
   GEMINI_MODEL,
-  Gemini,
-  GeminiSession,
-  GeminiVertexSession,
   type GoogleGeminiSessionOptions,
-  type VertexGeminiSessionOptions,
-} from "./gemini.js";
+} from "./gemini/types.js";
+
 export { Groq } from "./groq.js";
 export { HuggingFaceInferenceAPI } from "./huggingface.js";
 export {
